refactor(Sprint8): Page 및 layout 파일 경로 변경
